Separate Content file in CoursePage
