+++ src/components/Header/Header.tsx
@@ -5,7 +5,7 @@ import { faBell } from '@fortawesome/free-solid-svg-icons'
 import { useState } from 'react'
 import Noti from './Noti'
 import { useContext } from 'react'
-import { userContext } from '~/context/userContext'
+import { userContext } from '~/context/User/userContext.tsx'
 import type { MenuProps } from 'antd'
 import { Dropdown, Avatar } from 'antd'
 import { useNavigate } from 'react-router-dom'
