+++ src/pages/Auth/Login.tsx
@@ -8,11 +8,11 @@ import {
   PASSWORD_RULE_MESSAGE
 } from '~/utils/validators.ts'
 import FieldErrorAlert from '~/components/Form/FieldErrorAlert.tsx'
-import { userContext } from '~/context/userContext.tsx'
+import { userContext } from '~/context/User/userContext.tsx'
 import { toast } from 'react-toastify'
 import { userLoginAPI } from '~/apis/userAPI'
 // import { AxiosError } from 'axios'
-import { type User } from '~/context/userContext'
+import { type User } from '~/context/User/userContext.tsx'
 
 type LoginFormData = {
   username: string
