--- src/pages/Course/Document.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Document: React.FC =() => {
-  return (
-    <Document>
-
-    </Document>
-  )
-}
-
-export default Document
